+++ packages/blocks/src/page-block/default/selection-manager.ts
@@ -338,75 +338,6 @@ export class DefaultSelectionManager {
     this._mouseRoot = mouseRoot;
     this._container = container;
 
-    const createBlockHub = () => {
-      this._blockHub = new BlockHub({
-        onDropCallback: (e, end) => {
-          const dataTransfer = e.dataTransfer;
-          assertExists(dataTransfer);
-          const data = dataTransfer.getData('affine/block-hub');
-          const blockProps = JSON.parse(data);
-          if (blockProps.flavour === 'affine:database') {
-            if (!page.awarenessStore.getFlag('enable_database')) {
-              console.warn('database block is not enabled');
-              return;
-            }
-          }
-          const targetModel = end.model;
-          const rect = end.position;
-          this.page.captureSync();
-          const distanceToTop = Math.abs(rect.top - e.y);
-          const distanceToBottom = Math.abs(rect.bottom - e.y);
-          const id = this.page.addSiblingBlock(
-            targetModel,
-            blockProps,
-            distanceToTop < distanceToBottom ? 'right' : 'left'
-          );
-          asyncFocusRichText(this.page, id);
-        },
-        getBlockEditingStateByPosition: (blocks, pageX, pageY, skipX) => {
-          return getBlockEditingStateByPosition(blocks, pageX, pageY, {
-            skipX,
-          });
-        },
-        getBlockEditingStateByCursor: (
-          blocks,
-          pageX,
-          pageY,
-          cursor,
-          size,
-          skipX,
-          dragging
-        ) => {
-          return getBlockEditingStateByCursor(blocks, pageX, pageY, cursor, {
-            size,
-            skipX,
-            dragging,
-          });
-        },
-      });
-      this._blockHub.getAllowedBlocks = () => this._allowSelectedBlocks;
-    };
-    this._disposables.add(
-      this.page.awarenessStore.signals.update.subscribe(
-        msg => msg.state?.flags.enable_block_hub,
-        enable => {
-          if (enable) {
-            if (!this._blockHub) {
-              createBlockHub();
-            }
-          } else {
-            this._blockHub?.remove();
-            this._blockHub = null;
-          }
-        },
-        {
-          filter: msg => msg.id === this.page.doc.clientID,
-        }
-      )
-    );
-    if (this.page.awarenessStore.getFlag('enable_block_hub')) {
-      createBlockHub();
-    }
     this._embedResizeManager = new EmbedResizeManager(this.state, signals);
     this._disposables.add(
       initMouseEventHandlers(
