+++ packages/blocks/src/page-block/default/utils.ts
@@ -5,11 +5,12 @@ import {
 } from '@blocksuite/global/config';
 import { assertExists, matchFlavours } from '@blocksuite/global/utils';
 import type { BaseBlockModel } from '@blocksuite/store';
-import { DragHandle } from '../../components/index.js';
+import { BlockHub, DragHandle } from '../../components/index.js';
 import { toast } from '../../components/toast.js';
 import type { EmbedBlockModel } from '../../embed-block/embed-model.js';
 import { getService } from '../../__internal__/service.js';
 import {
+  asyncFocusRichText,
   doesInSamePath,
   getBlockById,
   getBlockElementByModel,
@@ -20,6 +21,7 @@ import type {
   CodeBlockOption,
   DefaultPageBlockComponent,
 } from './default-page-block.js';
+import { EdgelessPageBlockComponent } from '../edgeless/edgeless-page-block.js';
 
 export interface EditingState {
   model: BaseBlockModel;
@@ -467,4 +469,57 @@ export function createDragHandle(defaultPageBlock: DefaultPageBlockComponent) {
       );
     },
   });
+}
+
+export function createBlockHub(
+  pageBlock: DefaultPageBlockComponent | EdgelessPageBlockComponent
+) {
+  const page = pageBlock.page;
+  const model = pageBlock.model;
+  return new BlockHub({
+    onDropCallback: (e, end) => {
+      const dataTransfer = e.dataTransfer;
+      assertExists(dataTransfer);
+      const data = dataTransfer.getData('affine/block-hub');
+      const blockProps = JSON.parse(data);
+      if (blockProps.flavour === 'affine:database') {
+        if (!page.awarenessStore.getFlag('enable_database')) {
+          console.warn('database block is not enabled');
+          return;
+        }
+      }
+      const targetModel = end.model;
+      const rect = end.position;
+      page.captureSync();
+      const distanceToTop = Math.abs(rect.top - e.y);
+      const distanceToBottom = Math.abs(rect.bottom - e.y);
+      const id = page.addSiblingBlock(
+        targetModel,
+        blockProps,
+        distanceToTop < distanceToBottom ? 'right' : 'left'
+      );
+      asyncFocusRichText(page, id);
+    },
+    getBlockEditingStateByPosition: (blocks, pageX, pageY, skipX) => {
+      return getBlockEditingStateByPosition(blocks, pageX, pageY, {
+        skipX,
+      });
+    },
+    getBlockEditingStateByCursor: (
+      blocks,
+      pageX,
+      pageY,
+      cursor,
+      size,
+      skipX,
+      dragging
+    ) => {
+      return getBlockEditingStateByCursor(blocks, pageX, pageY, cursor, {
+        size,
+        skipX,
+        dragging,
+      });
+    },
+  });
+  this._blockHub.getAllowedBlocks = () => getAllowSelectedBlocks(model);
 }
